chore: process ui state on render
