+++ src/lib/sass/base/_mixins.scss
@@ -261,5 +261,4 @@
   overflow: hidden;
   box-shadow: none;
   margin: 0;
-  display: none;
 }
